18 ödev bitti :)

+++ Entities/DTOs/CarDetailDto.cs
@@ -12,7 +12,7 @@ namespace Entities.DTOs
         public string ColorName { get; set; }
         public string BrandName { get; set; }
 
-        public decimal DailyPrice { get; set; }
+        public int DailyPrice { get; set; }
         public int ModelYear { get; set; }
         public string Description { get; set; }
